pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "reddit-automation"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0",
  "praw>=7.7.1",
  "jsonschema>=4.23.0",
  "google-api-python-client>=2.120.0",
  "google-auth-httplib2>=0.2.0",
  "google-auth-oauthlib>=1.2.0",
  "google-auth>=2.29.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
